feat(navBubble) - refine item placement on bubble
